+++ Engine/core-engine/src/GameObject/transform-component.cpp
@@ -80,19 +80,19 @@ void TransformComponent::inspector_view()
             ImGui::TableNextColumn();
             ImGui::PushID(0);
             ImGui::Text("X"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &position.x);
+            ImGui::InputFloat("", &position.x);
             ImGui::PopID();
 
             ImGui::TableNextColumn();
             ImGui::PushID(1);
             ImGui::Text("Y"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &position.y);
+            ImGui::InputFloat("", &position.y);
             ImGui::PopID();
 
             ImGui::TableNextColumn();
             ImGui::PushID(2);
             ImGui::Text("Z"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &position.z);
+            ImGui::InputFloat("", &position.z);
             ImGui::PopID();
 
             ImGui::EndTable();
@@ -107,19 +107,19 @@ void TransformComponent::inspector_view()
             ImGui::TableNextColumn();
             ImGui::PushID(0);
             ImGui::Text("X"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &rotation.x);
+            ImGui::InputFloat("", &rotation.x);
             ImGui::PopID();
 
             ImGui::TableNextColumn();
             ImGui::PushID(1);
             ImGui::Text("Y"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &rotation.y);
+            ImGui::InputFloat("", &rotation.y);
             ImGui::PopID();
 
             ImGui::TableNextColumn();
             ImGui::PushID(2);
             ImGui::Text("Z"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &rotation.z);
+            ImGui::InputFloat("", &rotation.z);
             ImGui::PopID();
 
             ImGui::EndTable();
@@ -134,19 +134,19 @@ void TransformComponent::inspector_view()
             ImGui::TableNextColumn();
             ImGui::PushID(0);
             ImGui::Text("X"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &scale.x);
+            ImGui::InputFloat("", &scale.x);
             ImGui::PopID();
 
             ImGui::TableNextColumn();
             ImGui::PushID(1);
             ImGui::Text("Y"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &scale.y);
+            ImGui::InputFloat("", &scale.y);
             ImGui::PopID();
 
             ImGui::TableNextColumn();
             ImGui::PushID(2);
             ImGui::Text("Z"); ImGui::SameLine(); ImGui::SetNextItemWidth(-FLT_MIN);
-            ImGui::InputDouble("", &scale.z);
+            ImGui::InputFloat("", &scale.z);
             ImGui::PopID();
 
             ImGui::EndTable();
